+++ fixtures/_testPackage/web/main.dart
@@ -6,7 +6,9 @@
 // @dart=2.9
 
 import 'dart:async';
+import 'dart:convert';
 import 'dart:core';
+import 'dart:developer';
 import 'dart:html';
 
 import 'package:_test/deferred_library.dart' deferred as d;
@@ -30,6 +32,7 @@ void main() {
   });
 
   // for evaluation
+  var extensionId = 0;
   Timer.periodic(const Duration(seconds: 1), (_) {
     printLocal();
     printFieldFromLibraryClass(); // Breakpoint: callPrintFieldFromLibraryClass
@@ -47,9 +50,18 @@ void main() {
     printList();
     printMap();
     printSet();
+    // For testing evaluation in async JS frames.
+    registerUserExtension(extensionId++);
   });
 
   document.body.appendText(concatenate('Program', ' is running!'));
+}
+
+void registerUserExtension(int id) async {
+  registerExtension('ext.extension$id', (_, __) async {
+    print('Hello World from extension$id');
+    return ServiceExtensionResponse.result(json.encode({'success': true}));
+  });
 }
 
 void printLocal() {
@@ -68,7 +80,7 @@ void printFieldFromLibraryPartClass() {
 }
 
 void printFieldMain() {
-  var instance = MainClass(2,1);
+  var instance = MainClass(2, 1);
   print('$instance'); // Breakpoint: printFieldMain
 }